@@ -55,8 +55,7 @@ class StringStore {
 public:
 
     explicit StringStore(size_t chunk_size) :
-        m_chunk_size(chunk_size),
-        m_chunks() {
+        m_chunk_size(chunk_size) {
         add_chunk();
     }
 
